+++ MixItUp.WPF/Services/DeveloperAPI/V2/InventoryV2Controller.cs
@@ -3,7 +3,6 @@ using MixItUp.Base;
 using MixItUp.Base.Services;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 
@@ -39,57 +38,6 @@ namespace MixItUp.WPF.Services.DeveloperAPI.V2
             }
 
             return Ok(inventories);
-        }
-
-        [Route("{inventoryId:guid}/{itemId:guid}/top")]
-        [HttpGet]
-        public async Task<IActionResult> GetTopInventoryItemUsers(Guid inventoryId, Guid itemId, [FromQuery] int count = 10)
-        {
-            if (!ChannelSession.Settings.Inventory.TryGetValue(inventoryId, out var inventory) || inventory == null)
-            {
-                return NotFound(new ProblemDetails
-                {
-                    Status = 404,
-                    Title = "Not Found",
-                    Detail = $"Inventory with ID '{inventoryId}' not found"
-                });
-            }
-
-            var item = inventory.GetItem(itemId);
-            if (item == null)
-            {
-                return NotFound(new ProblemDetails
-                {
-                    Status = 404,
-                    Title = "Not Found",
-                    Detail = $"Item with ID '{itemId}' not found in inventory"
-                });
-            }
-
-            if (count <= 0) { count = 10; }
-
-            await ServiceManager.Get<UserService>().LoadAllUserData();
-
-            var topUsers = ChannelSession.Settings.Users.Values
-                .Where(u => u.InventoryAmounts.ContainsKey(inventoryId) && u.InventoryAmounts[inventoryId].ContainsKey(itemId) && u.InventoryAmounts[inventoryId][itemId] > 0)
-                .OrderByDescending(u => u.InventoryAmounts[inventoryId][itemId])
-                .Take(count)
-                .Select(u => new InventoryItemUserAmount
-                {
-                    UserID = u.ID,
-                    Username = u.GetAllPlatformUsernames().FirstOrDefault() ?? string.Empty,
-                    Amount = u.InventoryAmounts[inventoryId][itemId]
-                })
-                .ToList();
-
-            return Ok(new GetTopInventoryItemUsersResponse
-            {
-                InventoryID = inventoryId,
-                InventoryName = inventory.Name,
-                ItemID = itemId,
-                ItemName = item.Name,
-                Users = topUsers
-            });
         }
 
         [Route("{inventoryId:guid}/{userId:guid}")]
